Move crawl code to pkg/boot/crawl.
