+++ pkg/boot/crawl/beacon.go
@@ -1,4 +1,4 @@
-package boot
+package crawl
 
 import (
 	"bytes"
