+++ pkg/boot/crawl/strategy.go
@@ -1,4 +1,4 @@
-package boot
+package crawl
 
 import (
 	"context"
